UBUNTU: SAUCE: hid: ntrig: remove sysfs nodes
Aiming at a driver that requires no tuning, remove the sysfs nodes
for the filter parameters.

Has been tested succesfully on Dell Studio 17, Dell XT2, HP TX2 and
Lenovo T410s. Not likely to go upstream in its present form. However,
it does help constitue a tuning-free driver which works well together
with Unity, and as such are vital to the MT push.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>