yama: if an underlying filesystem provides a permissions op use it
When we are checking permissions on hardlinks we use generic_permissions()
to work out if the user actually has read/write permissions and only
then allow the link.  However where the underlying filesystem supplies
a permissions() op there is no guarentee that the inode ownership is
actually valid and we must use that op instead.

Add a new function mirroring the core fragment from inode_permission
using the filesystem specific permissions() op falling back to
generic_permissions() when it is not present.

With this in place links in overlayfs behave as expected.

Signed-off-by: Andy Whitcroft <apw@canonical.com>