UBUNTU: ubuntu: overlayfs -- ovl: fix overlayfs over overlayfs
Overlayfs expects ->permission() to not check for readonliness (which
is normally checked by the VFS) and so not return with -EROFS.  This
is not true of some filesystems, notably overlayfs itself.

The following patch should fix this by making sure that if the upper
layer is read-only (such as squashfs) then it will mark overlayfs
read-only too and by making ovl_permission() only return EROFS in the
excpetional case where the upper filesystem became r/o after the
overlay was constructed.

Reported-by: Jordi Pujol <jordipujolp@gmail.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Signed-off-by: Andy Whitcroft <apw@canonical.com>