UBUNTU: [Upstream] xen-blkfront: Drop name and minor adjustments for emulated scsi devices
[ This patch (xvd device naming) has been accepted by the Xen maintainer
but seems to be blocked somewhere in the block device maintainers queue
through which it has to go. ]

These were intended to avoid the namespace clash when representing
emulated IDE and SCSI devices. However that seems to confuse users
more than expected (a disk defined as sda becomes xvde).
So for now go back to the scheme which does no adjustments. This
will break when mixing IDE and SCSI names in the configuration of
guests but should be by now expected.

BugLink: http://bugs.launchpad.net/bugs/784937

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>