SGX: UDD: Avoid sending a black frame at initialization
This patch removes the fb_set_var used when the flipping is being done
with the DSS APIs. When the UDD initializes it asks the framebuffer
if it can hold certain number of buffers, that call makes a transfer to
the display of whatever contents are in the framebuffer (black at init
time).This patch avoids such transfer.

Change-Id: I04972ce1f9ac8b9b0f1a1232da01e1d68664dc53
Signed-off-by: sayen mohanty <sayenmohanty@ti.com>
Signed-off-by: Gustavo Diaz Prado <x0083741@ti.com>